refactor: util 파일 이름 변경
