+++ src/hooks/useCardNumber.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { toOnlyNumber, toHiddenNumber } from "../util/InputUtil";
+import { toOnlyNumber, toHiddenNumber } from "../util/replace";
 import { LENGTH, STRING } from "../abstract/constants";
 
 function useCardNumber() {
